+++ System.IO.Abstractions.TestingHelpers/MockDirectory.cs
@@ -117,7 +117,7 @@ namespace System.IO.Abstractions.TestingHelpers
             
             if (!mockFileDataAccessor.Directory.Exists(path))
             {
-                throw new DirectoryNotFoundException(string.Format(CultureInfo.InvariantCulture, StringResources.Manager.GetString("COULD_NOT_FIND_PART_OF_PATH_EXCEPTION"), path));
+                throw CommonExceptions.CouldNotFindPartOfPath(path);
             }
 
             var directoryData = (MockDirectoryData) mockFileDataAccessor.GetFile(path);
@@ -203,11 +203,7 @@ namespace System.IO.Abstractions.TestingHelpers
 
             if (!Exists(path))
             {
-                throw new DirectoryNotFoundException(
-                    string.Format(
-                        CultureInfo.InvariantCulture,
-                        StringResources.Manager.GetString("COULD_NOT_FIND_PART_OF_PATH_EXCEPTION"),
-                        path));
+                throw CommonExceptions.CouldNotFindPartOfPath(path);
             }
 
             path = EnsureAbsolutePath(path);
@@ -402,7 +398,7 @@ namespace System.IO.Abstractions.TestingHelpers
 
             if (!mockFileDataAccessor.Directory.Exists(path))
             {
-                throw new DirectoryNotFoundException(string.Format(CultureInfo.InvariantCulture, StringResources.Manager.GetString("COULD_NOT_FIND_PART_OF_PATH_EXCEPTION"), path));
+                throw CommonExceptions.CouldNotFindPartOfPath(path);
             }
 
             var directoryData = (MockDirectoryData)mockFileDataAccessor.GetFile(path);
